Split CacheObjectProvider and CacheObject in 2 files
