+++ test/cache_object_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
-import 'package:flutter_cache_manager/src/cache_object.dart';
+import 'package:flutter_cache_manager/src/storage/cache_object.dart';
+import 'package:flutter_cache_manager/src/storage/cache_object_provider.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path/path.dart' as p;
